refactor(wallet-connect): consolidate domain into src/features/wallet-connect
