+++ src/features/wallet-connect/components/WalletConnectV2ListItem.tsx
@@ -11,17 +11,17 @@ import styled from '@/framework/ui/styled-thing';
 import { padding, position } from '@/styles';
 import { useTheme } from '@/theme';
 import { showActionSheetWithOptions } from '@/framework/ui/utils/actionsheet';
-import { changeAccount, disconnectSession } from '@/walletConnect';
+import { changeAccount, disconnectSession } from '..';
 import { type SessionTypes } from '@walletconnect/types';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import RadialGradient from 'react-native-radial-gradient';
 import { getAccountProfileInfo } from '@/state/wallets/walletsStore';
-import RequestVendorLogoIcon from '../coin-icon/RequestVendorLogoIcon';
-import { ContactAvatar } from '../contacts';
-import ImageAvatar from '../contacts/ImageAvatar';
-import { ContextMenuButton } from '../context-menu';
-import { Centered, ColumnWithMargins, Row } from '../layout';
-import { TruncatedText } from '../text';
+import RequestVendorLogoIcon from '@/components/coin-icon/RequestVendorLogoIcon';
+import { ContactAvatar } from '@/components/contacts';
+import ImageAvatar from '@/components/contacts/ImageAvatar';
+import { ContextMenuButton } from '@/components/context-menu';
+import { Centered, ColumnWithMargins, Row } from '@/components/layout';
+import { TruncatedText } from '@/components/text';
 import { isValidHex } from '@/handlers/web3';
 import { opacity } from '@/framework/ui/utils/opacity';
