+++ src/hooks/useApplicationSetup.ts
@@ -6,7 +6,7 @@ import { type InitialRoute } from '@/navigation/initialRoute';
 import Routes from '@/navigation/routesNames';
 import { saveFCMToken } from '@/notifications/tokens';
 import { PerformanceReports, PerformanceTracking } from '@/performance/tracking';
-import { initListeners as initWalletConnectListeners, initWalletConnectPushNotifications } from '@/walletConnect';
+import { initListeners as initWalletConnectListeners, initWalletConnectPushNotifications } from '@/features/wallet-connect';
 import { initializeWallet } from '@/state/wallets/initializeWallet';
 
 export function useApplicationSetup() {
